Feature/43 introduce appointment response/request and appointment controller and service layer also Tests (#61)
* 45: check if Appointment is in the past

* 45: implemented findAppointmentsByAuthor and deleteAppointment logic

* 43: added updateAppointment in controller and logic in service

* 43: refactoring controller

* 43: AppointmentRequest/Mapper/Controller Tests added

* 43: update tests AppointmentController and refactor

* 43: Tests for appointmentService and some refactor

* 43: change argument naming of AppointmentResponse author to email

* 43: change HttpStatus code in Test

* 43: finish Appointment Tests and refactor

---------

+++ src/main/java/com/calendar/services/AppointmentService.java
@@ -6,10 +6,11 @@ import com.calendar.entities.User;
 import com.calendar.mapper.AppointmentMapper;
 import com.calendar.repositories.IAppointmentRepository;
 import com.calendar.repositories.IUserRepository;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import jakarta.annotation.Nullable;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -18,38 +19,62 @@ public class AppointmentService {
     private final AppointmentMapper appointmentMapper;
     private final IUserRepository userRepository;
 
-    public AppointmentService(final IAppointmentRepository appointmentRepository, final AppointmentMapper appointmentMapper, final IUserRepository userRepository) {
+    public AppointmentService(final IAppointmentRepository appointmentRepository,
+                              final AppointmentMapper appointmentMapper,
+                              final IUserRepository userRepository) {
         this.appointmentRepository = appointmentRepository;
         this.appointmentMapper = appointmentMapper;
         this.userRepository = userRepository;
     }
 
-    public ResponseEntity<String> create(final AppointmentRequest appointmentRequest) {
-        final User user = userRepository.findByEmail(appointmentRequest.author());
+    public boolean create(final AppointmentRequest appointmentRequest) {
+
+        final User user = userRepository.findByEmail(appointmentRequest.email());
         if (user == null)
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cannot create Appointment, User could not be found!");
+            return false;
 
         Appointment appointment = appointmentMapper.toAppointment(appointmentRequest, user);
         appointmentRepository.save(appointment);
-        return ResponseEntity.ok("Appointment successfully created!");
+        return true;
     }
 
-    public void delete(final Appointment appointment) {
-        appointmentRepository.delete(appointment);
-    }
+    public boolean deleteById(final UUID id) {
 
-    public void deleteById(final UUID id) {
+        final Optional<Appointment> appointment = appointmentRepository.findById(id);
+        if (appointment.isEmpty())
+            return false;
         appointmentRepository.deleteById(id);
+        return true;
     }
 
-    public void update(final Appointment appointment) {
+    public boolean update(final AppointmentRequest appointmentRequest) {
+
+        //Get optionalAppointment from repo
+        final Optional<Appointment> optionalAppointment = appointmentRepository.findById(UUID.
+                fromString(appointmentRequest.id()));
+
+        // check if present and same author
+        if (optionalAppointment.isEmpty() || !appointmentRequest.email().equals(optionalAppointment.get().getAuthor().
+                getEmail()))
+            return false;
+
+        Appointment appointment = optionalAppointment.get();
+        appointment.setTitle(appointmentRequest.title());
+        appointment.setDescription(appointmentRequest.description());
+        appointment.setStartDateTime(appointmentRequest.startDateTime());
+        appointment.setEndDateTime(appointmentRequest.endDateTime());
+
         appointmentRepository.save(appointment);
+        return true;
     }
 
-    //TODO
-//    public List<Appointment> findByAuthor(final UserDTO authorDto) {
-//        User author = new User();
-//        return appointmentRepository.findByAuthor(author);
-//    }
+    @Nullable
+    public List<Appointment> findByEmail(final String email) {
 
+        User user = userRepository.findByEmail(email);
+
+        if (user == null)
+            return null;
+        return appointmentRepository.findByAuthor(user);
+    }
 }
